requirements.txt
aniso8601==9.0.1
attrs==21.2.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.6
click==8.0.1
colorama==0.4.4
coloredlogs==15.0.1
dnspython==2.1.0
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.0
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
flask-restx==0.5.1
gunicorn==20.1.0
humanfriendly==9.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.1.0
pymongo==3.12.0
pyreadline==2.1
pyrsistent==0.18.0
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
Werkzeug==2.0.1
